Renamed ModifyMoveset to ModifyMovesetModal
